# Avoid recomputing bindings multiple times and bind BindingContext first