--- a/src/Controls/src/Core/BindableObject.cs
+++ b/src/Controls/src/Core/BindableObject.cs
@@ -341,7 +341,7 @@ namespace Microsoft.Maui.Controls
 		[EditorBrowsable(EditorBrowsableState.Never)]
 		public static void SetInheritedBindingContext(BindableObject bindable, object value)
 		{
-			//I wonder if we coulnd't treat bindingcoutext with specificities
+			// I wonder if we couldn't treat BindingContext with specificities
 			BindablePropertyContext bpContext = bindable.GetContext(BindingContextProperty);
 			if (bpContext != null && bpContext.Values.GetSpecificityAndValue().Key.CompareTo(SetterSpecificity.ManualValueSetter) >= 0)
 				return;
@@ -355,20 +355,33 @@ namespace Microsoft.Maui.Controls
 			{
 				binding.Context = value;
 				bindable._inheritedContext = null;
+				// OnBindingContextChanged fires from within BindingContextProperty propertyChanged callback
+				bindable.ApplyBinding(bpContext, fromBindingContextChanged: true);
 			}
 			else
 			{
 				bindable._inheritedContext = new WeakReference(value);
+				bindable.ApplyBindings(fromBindingContextChanged: true);
+				bindable.OnBindingContextChanged();
 			}
-
-			bindable.ApplyBindings(skipBindingContext: false, fromBindingContextChanged: true);
-			bindable.OnBindingContextChanged();
 		}
 
 		/// <summary>
 		/// Applies all the current bindings to <see cref="BindingContext" />.
 		/// </summary>
-		protected void ApplyBindings() => ApplyBindings(skipBindingContext: false, fromBindingContextChanged: false);
+		protected void ApplyBindings()
+		{
+			BindablePropertyContext bpContext = GetContext(BindingContextProperty);
+			var binding = bpContext?.Bindings.Values.LastOrDefault();
+			if (binding != null)
+			{
+				ApplyBinding(bpContext, fromBindingContextChanged: false);
+			}
+			else
+			{
+				ApplyBindings(fromBindingContextChanged: false);
+			}
+		}
 
 		/// <summary>
 		/// Raises the <see cref="BindingContextChanged"/> event.
@@ -645,25 +658,43 @@ namespace Microsoft.Maui.Controls
 			}
 		}
 
-		internal void ApplyBindings(bool skipBindingContext, bool fromBindingContextChanged)
+		void ApplyBindings(bool fromBindingContextChanged)
 		{
 			var prop = _properties.Values.ToArray();
+
 			for (int i = 0, propLength = prop.Length; i < propLength; i++)
 			{
 				BindablePropertyContext context = prop[i];
-				var kvp = context.Bindings.LastOrDefault();
-				var specificity = kvp.Key;
-				var binding = kvp.Value;
-
-				if (binding == null)
+				if (ReferenceEquals(context.Property, BindingContextProperty))
+				{
+					// BindingContextProperty Binding is handled separately within SetInheritedBindingContext
 					continue;
+				}
 
-				if (skipBindingContext && ReferenceEquals(context.Property, BindingContextProperty))
-					continue;
-
-				binding.Unapply(fromBindingContextChanged: fromBindingContextChanged);
-				binding.Apply(BindingContext, this, context.Property, fromBindingContextChanged, specificity);
+				ApplyBinding(context, fromBindingContextChanged);
 			}
+		}
+
+		[MethodImpl(MethodImplOptions.AggressiveInlining)]
+		void ApplyBinding(BindablePropertyContext context, bool fromBindingContextChanged)
+		{
+			var bindings = context.Bindings;
+			if (bindings.Count == 0)
+			{
+				return;
+			}
+
+			var kvp = bindings.LastOrDefault();
+			var binding = kvp.Value;
+
+			if (binding == null)
+			{
+				return;
+			}
+
+			var specificity = kvp.Key;
+			binding.Unapply(fromBindingContextChanged);
+			binding.Apply(BindingContext, this, context.Property, fromBindingContextChanged, specificity);
 		}
 
 		static void BindingContextPropertyBindingChanging(BindableObject bindable, BindingBase oldBindingBase, BindingBase newBindingBase)
@@ -681,7 +712,7 @@ namespace Microsoft.Maui.Controls
 		static void BindingContextPropertyChanged(BindableObject bindable, object oldvalue, object newvalue)
 		{
 			bindable._inheritedContext = null;
-			bindable.ApplyBindings(skipBindingContext: true, fromBindingContextChanged: true);
+			bindable.ApplyBindings(fromBindingContextChanged: true);
 			bindable.OnBindingContextChanged();
 		}