### Description of Change

The two unit tests shows the expectation we all have about bindings:
- eventual `BindingContext` binding should apply before any property binding
- `Binding` should happen only once upon binding context inheritance

### Benchmarks

#### Before

| Method                                          | Mean     | Error   | StdDev  | Gen0     | Gen1    | Gen2    | Allocated |
|------------------------------------------------ |---------:|--------:|--------:|---------:|--------:|--------:|----------:|
| SetProperty                                     | 892.2 us | 7.64 us | 6.77 us |        - |       - |       - |     873 B |
| ApplyProperty ViaBindingContextInheritance       | 379.7 us | 3.94 us | 3.49 us |  76.6602 | 38.0859 |  3.9063 |  644003 B |
| ApplyBindingContext ViaBindingContextInheritance | 775.7 us | 8.05 us | 7.13 us | 109.3750 | 54.6875 | 11.7188 |  919289 B |

#### After

| Method                                          | Mean     | Error   | StdDev  | Gen0    | Gen1    | Allocated |
|------------------------------------------------ |---------:|--------:|--------:|--------:|--------:|----------:|
| SetProperty                                     | 868.0 us | 1.00 us | 0.83 us |       - |       - |     873 B |
| ApplyProperty ViaBindingContextInheritance       | 309.0 us | 0.54 us | 0.50 us | 72.7539 | 36.1328 |  612000 B |
| ApplyBindingContext ViaBindingContextInheritance | 497.6 us | 1.23 us | 1.15 us | 93.7500 | 46.8750 |  784881 B |

### Issues Fixed

Fixes #10806