@rmarinho 2024-09-13
@albyrock87  @StephaneDelcroix  this test seem to be failing when merging main to net9.0 https://dev.azure.com/xamarin/public/_build/results?buildId=123788&view=ms.vss-test-web.build-test-results-tab&runId=3145016&resultId=102857&paneView=debug