@albyrock87 2024-09-13
@rmarinho looking at your PR, I see a lot of changes in `BindableObject` are missing.
https://github.com/dotnet/maui/pull/24553/files#diff-ebb14ec14b8baeffdd7b0ae40ffc8bf0e8944f76fdea50fcbc737fb142bfb44a
Maybe a mistake solving merge conflicts?